// src/app/shared-modules/models/app-alert.interface.ts
export interface alertInfo {
    tipo: string;
    mensaje: string;
    visible: boolean;
  }
  export interface AlertInfo {
    type: string
    typeMsg?: string
    message: string
    visible: boolean
  }
  export class AlertInfo {
    type: string
    typeMsg?: string
    message: string
    visible: boolean
  }